pyproject.toml
[project]
name = "hrp"
version = "1.1.0"
description = "Hedgefund Research Platform - Personal quantitative research for systematic trading"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Fernando", email = "fmag79@gmail.com"}
]
keywords = ["quant", "trading", "research", "backtesting", "finance"]

dependencies = [
    # Data
    "duckdb>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyarrow>=14.0.0",

    # Backtesting
    "vectorbt>=0.26.0",

    # ML
    "scikit-learn>=1.3.0",
    "lightgbm>=4.0.0",
    "xgboost>=2.0.0",
    "optuna>=3.4.0",

    # Experiment tracking
    "mlflow>=2.10.0",

    # Dashboard
    "streamlit>=1.30.0",
    "plotly>=5.18.0",

    # Agent integration
    "fastmcp>=0.1.0",
    "anthropic>=0.18.0",

    # Scheduling & notifications
    "apscheduler>=3.10.0",
    "resend>=0.7.0",

    # Data sources
    "yfinance>=0.2.30",
    "polygon-api-client>=1.12.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "loguru>=0.7.0",
    "httpx>=0.26.0",
    "exchange-calendars>=4.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

polygon = [
    "polygon-api-client>=1.12.0",
]

alpaca = [
    "alpaca-py>=0.13.0",
]

[project.scripts]
hrp = "hrp.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=hrp --cov-report=term-missing"
